+++ src/frontend/components/PublicVideoDashboard/index.tsx
@@ -1,13 +1,21 @@
 import { Box } from 'grommet';
 import React from 'react';
+import { Redirect } from 'react-router-dom';
 
 import { Chat } from '../Chat';
 import { DownloadVideo } from '../DownloadVideo';
+import { FULL_SCREEN_ERROR_ROUTE } from '../ErrorComponents/route';
 import { Transcripts } from '../Transcripts';
 import VideoPlayer from '../VideoPlayer';
+import { WaitingLiveVideo } from '../WaitingLiveVideo';
 import { useTimedTextTrack } from '../../data/stores/useTimedTextTrack';
 import { useVideo } from '../../data/stores/useVideo';
-import { timedTextMode, TimedTextTranscript, Video } from '../../types/tracks';
+import {
+  liveState,
+  timedTextMode,
+  TimedTextTranscript,
+  Video,
+} from '../../types/tracks';
 
 interface PublicVideoDashboardProps {
   video: Video;
@@ -23,21 +31,33 @@ const PublicVideoDashboard = ({
     state.getTimedTextTracks(),
   );
 
-  if (video.live_state !== null && video.xmpp) {
-    return (
-      <Box direction="row">
-        <Box basis="xlarge">
-          <VideoPlayer
-            video={video}
-            playerType={playerType}
-            timedTextTracks={[]}
-          />
-        </Box>
-        <Box>
-          <Chat video={video} />
-        </Box>
-      </Box>
-    );
+  if (video.live_state !== null) {
+    switch (video.live_state) {
+      case liveState.RUNNING:
+        return (
+          <Box direction="row">
+            <Box basis={video.xmpp ? 'xlarge' : 'auto'}>
+              <VideoPlayer
+                video={video}
+                playerType={playerType}
+                timedTextTracks={[]}
+              />
+            </Box>
+            {video.xmpp && (
+              <Box>
+                <Chat video={video} />
+              </Box>
+            )}
+          </Box>
+        );
+      case liveState.STOPPED:
+      case liveState.STOPPING:
+        // nothing to show
+        return <Redirect push to={FULL_SCREEN_ERROR_ROUTE('notFound')} />;
+      default:
+        // waiting message
+        return <WaitingLiveVideo video={video} />;
+    }
   }
 
   const transcripts = timedTextTracks