+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Move DevelopmentLTIView to a new development directory
 - Add a variable to set frontend video polling interval
 - Add a component responsible to manage public video dashboard
+- Add a waiting workflow while a live is not ready
 
 ### Changed